--- benchmarks/perception/a8_stereo_image_proc_pc/launch/stereo_image_test.launch.py
@@ -1,56 +1,0 @@
-#
-#    @@@@@@@@@@@@@@@@@@@@
-#    @@@@@@@@@&@@@&&@@@@@
-#    @@@@@ @@  @@    @@@@
-#    @@@@@ @@  @@    @@@@
-#    @@@@@ @@  @@    @@@@ Copyright (c) 2023, Acceleration Robotics®
-#    @@@@@ @@  @@    @@@@
-#    @@@@@ @@  @@    @@@@
-#    @@@@@@@@@&@@@@@@@@@@
-#    @@@@@@@@@@@@@@@@@@@@
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import os
-from launch import LaunchDescription
-from launch_ros.actions import ComposableNodeContainer
-from launch_ros.descriptions import ComposableNode
-from launch_ros.actions import Node
-
-def generate_launch_description():
-    ld = LaunchDescription()
-
-    container = ComposableNodeContainer(
-        name='stereo_image_proc_container',
-        namespace='',
-        package='rclcpp_components',
-        executable='component_container',
-        composable_node_descriptions=[
-            ComposableNode(
-                package='stereo_image_proc',
-                plugin='stereo_image_proc::DisparityNode',
-                name='stereo_image_proc_disparity_node',
-                remappings=[
-                    ('left/camera_info', '/left_camera/camera_info'),
-                    ('left/image_rect', '/left_camera/image_raw'),
-                    ('right/camera_info', '/right_camera/camera_info'),
-                    ('right/image_rect', '/right_camera/image_raw'),                    
-                ],
-            )
-        ],
-        output='screen',
-    )
-
-    ld.add_action(container)
-
-    return ld
